% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.r
\name{add_label<-}
\alias{add_label<-}
\title{add a label (description) to an object (can be the column of a data.frame)}
\usage{
add_label(x) <- value
}
\arguments{
\item{x}{an object (can be the column of a data.frame)}

\item{value}{the label of the object (a character vector of length 1).}
}
\description{
add a label (description) to an object (can be the column of a data.frame)
}
\author{
David Hajage
}
