% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{body_add_crosstable}
\alias{body_add_crosstable}
\title{add a table made by the cross function into a officer document}
\usage{
body_add_crosstable(doc, crosstable, compact = FALSE, id = ".id",
  variable = "variable", value = "value", effect = "effect", p = "p",
  w = NULL, pos = "after")
}
\arguments{
\item{doc}{a \code{docx} object created by \code{read_docx} function (package \code{flextable})}

\item{crosstable}{the result of \code{cross} function}

\item{compact}{compact the table?}

\item{id}{name of the 'id' column}

\item{variable}{name of the 'variable' column}

\item{value}{name of the 'value' column}

\item{effect}{name of the 'effect' column}

\item{p}{name of the 'p' column}

\item{w}{width of the table (if \code{NULL}, width of the current document is used)}

\item{pos}{where to add the flextable relative to the cursor, one of "after", "before", "on" (end of line) (see \code{?body_add_flextable}).}
}
\value{
A \code{docx} object
}
\description{
add a table made by the cross function into a officer document
}
\examples{
\dontrun{
library(biostat2)
library(officer)
library(flextable)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
doc <- read_docx()
doc <- body_add_crosstable(doc, mytable)
doc <- body_add_break(doc)
doc <- body_add_crosstable(doc, mytable, TRUE)
}
}
\author{
David Hajage
}
\keyword{univar}
