% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.r
\name{body_add_img_autofit}
\alias{body_add_img_autofit}
\title{add an image into an rdocx object, with autofiting of width and height}
\usage{
body_add_img_autofit(x, src, style = NULL, pos = "after",
  landscape = FALSE)
}
\arguments{
\item{x}{the rdocx object (created with the read_docx function of officer package)}

\item{src}{image filename, the basename of the file must not contain any blank}

\item{style}{paragraph style}

\item{pos}{where to add the new element relative to the cursor, one of "after", "before", "on"}

\item{landscape}{is the image inserted in a landscape section? (default: FALSE)}
}
\value{
a new rdocx object
}
\description{
add an image into an rdocx object, with autofiting of width and height
}
\examples{
\dontrun{
library(biostat2)
library(officer)

jpeg("carre.jpg", width = 300, height = 300)
plot(1)
dev.off()
jpeg("long.jpg", width = 300, height = 600)
plot(1)
dev.off()
jpeg("large.jpg", width = 600, height = 300)
plot(1)
dev.off()

doc <- read_docx()
doc <- body_add_img_autofit(doc, "carre.jpg")
doc <- body_add_img_autofit(doc, "long.jpg") 
doc <- body_add_img_autofit(doc, "large.jpg") 
doc <- body_end_section_portrait(doc)

doc <- body_add_img_autofit(doc, "carre.jpg", landscape = TRUE) 
doc <- body_add_img_autofit(doc, "long.jpg", landscape = TRUE) 
doc <- body_add_img_autofit(doc, "large.jpg", landscape = TRUE) 
doc <- body_end_section_landscape(doc)

print(doc, target = "toto.docx")
}
}
\author{
David Hajage
}
\keyword{univar}
