% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.r
\name{body_add_tableref}
\alias{body_add_tableref}
\title{Add a new paragraph with a tablereference2 style}
\usage{
body_add_tableref(x, value, depth = 1, style = "tablereference2")
}
\arguments{
\item{x}{the rdocx object (created with the read_docx function of officer package)}

\item{value}{a character string}

\item{depth}{integer, what is the depth of the current section to print}

\item{style}{a character string}
}
\value{
a new rdocx object
}
\description{
Add a new paragraph with a tablereference2 style
}
\examples{
\dontrun{
library(biostat2)
library(officer)
doc <- body_add_plotlegend2(doc, "Coucou")
}
}
\author{
David Hajage
}
\keyword{univar}
