% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer.r
\name{body_add_title}
\alias{body_add_title}
\title{Add a title}
\usage{
body_add_title(x, value, level = 1, style = "heading", pos = "after")
}
\arguments{
\item{x}{the rdocx object (created with the read_docx function of officer package)}

\item{value}{a character string}

\item{level}{a integer (1 to 9)}

\item{style}{a character string}

\item{pos}{where to add the new element relative to the cursor, one of "after", "before", "on"}
}
\value{
a new rdocx object
}
\description{
Add a title
}
\examples{
\dontrun{
library(biostat2)
library(officer)
doc <- body_add_itemize(doc, "Coucou")
}
}
\author{
David Hajage
}
\keyword{univar}
