% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.r
\name{effect.diff.mean.auto}
\alias{effect.diff.mean.auto}
\title{Effect measure for association between one continuous and one categorical variable}
\usage{
effect.diff.mean.auto(x, g, conf.level = 0.95, R = 500)
}
\arguments{
\item{x}{vector (the continuous variable)}

\item{g}{another vector (the grouping variable)}

\item{conf.level}{confidence interval level}

\item{R}{number of bootstrap replication}
}
\value{
a list with five componments
}
\description{
Effect measure for association between one continuous and one categorical variable
}
