% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{factory}
\alias{factory}
\title{Create functions that capture their values, errors, and warnings}
\usage{
factory(fun)
}
\arguments{
\item{fun}{A function}
}
\value{
A function
}
\description{
Convert plain old functions into functions that capture their values, errors, and warnings. Comes from: https://stackoverflow.com/questions/4948361/how-do-i-save-warnings-and-errors-as-output-from-a-function
}
\author{
David Hajage
}
