% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{flexcrosstable-flextable}
\alias{flexcrosstable-flextable}
\alias{flexcrosstable}
\title{Create a flextable object from a table made by the cross function}
\usage{
flexcrosstable(crosstable, compact = FALSE, id = ".id",
  variable = "variable", value = "value", effect = "effect",
  p = "p")
}
\arguments{
\item{crosstable}{the result of \code{cross} function}

\item{compact}{compact the table?}

\item{id}{name of the 'id' column}

\item{variable}{name of the 'variable' column}

\item{value}{name of the 'value' column}

\item{effect}{name of the 'effect' column}

\item{p}{name of the 'p' column}
}
\value{
A \code{flextable} object (see \code{flextable} package)
}
\description{
Create a flextable object from a table made by the cross function
}
\examples{
\dontrun{
library(biostat2)
library(flextable)
mytable <- cross(cbind(...) ~ tobgp, esoph, test = TRUE)
flexcrosstable(mytable)
flexcrosstable(mytable, TRUE)
}
}
\author{
David Hajage
}
\keyword{univar}
