% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.r
\name{get_label}
\alias{get_label}
\title{get a label (description) to an object (can be the column of a data.frame)}
\usage{
get_label(x)
}
\arguments{
\item{x}{an object (can be the column of a data.frame)}
}
\description{
get a label (description) to an object (can be the column of a data.frame)
}
\author{
David Hajage
}
