% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues.r
\name{mapvalues}
\alias{mapvalues}
\title{Extrated from the plyr package to no more depend on it. Sorry Hadley.}
\usage{
mapvalues(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{x}{x}

\item{from}{from}

\item{to}{to}

\item{warn_missing}{warn_missing}
}
\description{
Extrated from the plyr package to no more depend on it. Sorry Hadley.
}
\author{
David Hajage
}
\keyword{internal}
