% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{merge_at_custom}
\alias{merge_at_custom}
\title{merge_at_custom}
\usage{
merge_at_custom(ft, j, runs)
}
\arguments{
\item{ft}{ft}

\item{j}{j}

\item{runs}{runs}
}
\description{
merge_at_custom
}
\author{
David Hajage
}
\keyword{internal}
