% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{myfisher.test}
\alias{myfisher.test}
\title{A fisher test robust to exceeded memory}
\usage{
myfisher.test(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{another vector}
}
\value{
a htest object
}
\description{
Compute a fisher test as appropriate (with simulated p-value if necessary)
}
\author{
David Hajage
}
