% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCrossTable.r
\name{theme_cabane}
\alias{theme_cabane}
\title{Apply the cabane theme (blue header)}
\usage{
theme_cabane(x)
}
\arguments{
\item{x}{a flextable object}
}
\value{
A \code{flextable object}
}
\description{
Apply the cabane theme (blue header)
}
\examples{
\dontrun{
library(flextable)
theme_cabane(flextable(head(iris)))
}
}
\author{
David Hajage
}
\keyword{univar}
